FIX : cleaning code
